requirements.txt
Flask==2.3.2
gunicorn
openai==1.97.0
python-dotenv
